Move slack component creation from bus class to optimization_model.py
